web/requirements.txt
asgiref==3.7.2
Django==4.2
djangorestframework==3.14.0
python-decouple==3.8
python-slugify==8.0.1
pytz==2023.3.post1
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
psycopg2-binary==2.9.6